requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2==2.9.9
python-dotenv==1.0.1
pillow==10.4.0
django-filter~=24.2
djangorestframework-simplejwt==5.3.1
stripe~=10.4.0
celery~=5.4.0
redis==5.0.8
